requirements.txt
streamlit

python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
sqlalchemy


 